+++ pyobjc-core/PyObjCTest/test_hidden_selector.py
@@ -150,56 +150,91 @@ class TestHiddenSelector(TestCase):
         o = OCTestHidden.alloc().init()
 
         # Instance method
-        with self.assertRaisesRegex(
-            AttributeError, "'OCTestHidden' object has no attribute 'method'"
-        ):
-            o.method()
+        with self.subTest("instance method resolve normally"):
+            with self.assertRaisesRegex(
+                AttributeError, "'OCTestHidden' object has no attribute 'method'"
+            ):
+                o.method()
 
-        v = o.performSelector_(b"method")
-        self.assertEqual(v, 42)
+        with self.subTest("instance method with performSelector"):
+            v = o.performSelector_(b"method")
+            self.assertEqual(v, 42)
 
-        v = o.pyobjc_instanceMethods.method()
-        self.assertEqual(v, 42)
+        with self.subTest("instance method with methodForSelector"):
+            v = o.methodForSelector_(b"method")(o)
+            self.assertEqual(v, 42)
 
-        # Class method
-        with self.assertRaisesRegex(AttributeError, "clsmethod"):
-            OCTestHidden.clsmethod()
+        with self.subTest("instance method with instanceMethodForSelector"):
+            v = OCTestHidden.instanceMethodForSelector_(b"method")(o)
+            self.assertEqual(v, 42)
 
-        with self.assertRaisesRegex(AttributeError, "clsmethod2"):
-            OCTestHidden.clsmethod2()
+        with self.subTest("instance method with pyobjc_instanceMethods"):
+            v = o.pyobjc_instanceMethods.method()
+            self.assertEqual(v, 42)
 
-        with self.assertRaisesRegex(AttributeError, "anotherclsmethod"):
-            OCTestHidden.anotherclsmethod()
+        with self.subTest("instance method with pyobjc_instanceMethods on class"):
+            v = OCTestHidden.pyobjc_instanceMethods.method(o)
+            self.assertEqual(v, 42)
 
-        cls_dict = OCTestHidden.__dict__
-        self.assertNotIn("method", cls_dict)
-        self.assertNotIn("clsmethod", cls_dict)
-        self.assertNotIn("clsmethod2", cls_dict)
-        self.assertNotIn("anotherclsmethod", cls_dict)
+        with self.subTest("class method with pyobjc_classMethods"):
+            v = OCTestHidden.pyobjc_classMethods.clsmethod()
+            self.assertEqual(v, 99)
 
-        cls_dir = dir(OCTestHidden)
-        self.assertNotIn("method", cls_dir)
-        self.assertNotIn("clsmethod", cls_dir)
-        self.assertNotIn("clsmethod2", cls_dir)
-        self.assertNotIn("anotherclsmethod", cls_dir)
+        with self.subTest("class method with methodForSelector"):
+            v = OCTestHidden.methodForSelector_(b"clsmethod")(OCTestHidden)
+            self.assertEqual(v, 99)
 
-        cls_dir = dir(type(OCTestHidden))
-        self.assertNotIn("method", cls_dir)
-        self.assertNotIn("clsmethod", cls_dir)
-        self.assertNotIn("clsmethod2", cls_dir)
-        self.assertNotIn("anotherclsmethod", cls_dir)
+        with self.subTest("class method resolve normall"):
+            with self.assertRaisesRegex(AttributeError, "clsmethod"):
+                OCTestHidden.clsmethod()
 
-        v = OCTestHidden.performSelector_(b"clsmethod")
-        self.assertEqual(v, 99)
+            with self.assertRaisesRegex(AttributeError, "clsmethod2"):
+                OCTestHidden.clsmethod2()
 
-        v = OCTestHidden.performSelector_(b"anotherclsmethod")
-        self.assertEqual(v, -99)
+            with self.assertRaisesRegex(AttributeError, "anotherclsmethod"):
+                OCTestHidden.anotherclsmethod()
 
-        v = OCTestHidden.pyobjc_classMethods.clsmethod()
-        self.assertEqual(v, 99)
+        with self.subTest("validate class __dict__"):
+            cls_dict = OCTestHidden.__dict__
+            self.assertNotIn("method", cls_dict)
+            self.assertNotIn("clsmethod", cls_dict)
+            self.assertNotIn("clsmethod2", cls_dict)
+            self.assertNotIn("anotherclsmethod", cls_dict)
 
-        v = OCTestHidden.pyobjc_classMethods.anotherclsmethod()
-        self.assertEqual(v, -99)
+        with self.subTest("validate meta class __dict__"):
+            cls_dict = type(OCTestHidden).__dict__
+            self.assertNotIn("method", cls_dict)
+            self.assertNotIn("clsmethod", cls_dict)
+            self.assertNotIn("clsmethod2", cls_dict)
+            self.assertNotIn("anotherclsmethod", cls_dict)
+
+        with self.subTest("validate class dir()"):
+            cls_dir = dir(OCTestHidden)
+            self.assertNotIn("method", cls_dir)
+            self.assertNotIn("clsmethod", cls_dir)
+            self.assertNotIn("clsmethod2", cls_dir)
+            self.assertNotIn("anotherclsmethod", cls_dir)
+
+        with self.subTest("validate meta class __dict__"):
+            cls_dir = dir(type(OCTestHidden))
+            self.assertNotIn("method", cls_dir)
+            self.assertNotIn("clsmethod", cls_dir)
+            self.assertNotIn("clsmethod2", cls_dir)
+            self.assertNotIn("anotherclsmethod", cls_dir)
+
+        with self.subTest("class method with performSelector"):
+            v = OCTestHidden.performSelector_(b"clsmethod")
+            self.assertEqual(v, 99)
+
+            v = OCTestHidden.performSelector_(b"anotherclsmethod")
+            self.assertEqual(v, -99)
+
+        with self.subTest("class method with pyobjc_classMethods"):
+            v = OCTestHidden.pyobjc_classMethods.clsmethod()
+            self.assertEqual(v, 99)
+
+            v = OCTestHidden.pyobjc_classMethods.anotherclsmethod()
+            self.assertEqual(v, -99)
 
     def test_hidden_flag_invalid(self):
         @objc.selector